Update README.md
Added Line 4 to describe how you can read this README with each line numbered for easier reading. This is in-case you are reading this, but not entirely familiar with GitHub or your brain works such that you would prefer to be told that seeing the Line Numbers makes this README easier to understand.

I wouldn't want someone coming across the README simply needing this tool, but feeling lost with the instructions because I mentioned Line Numbers and you don't see any Line Numbers.